Move test for`rooney_biegler_doe_example.py` to `test_doe_solve.py`
