+++ pyomo/contrib/parmest/tests/test_examples.py
@@ -14,14 +14,9 @@ import pyomo.contrib.parmest.parmest as parmest
 from pyomo.contrib.parmest.graphics import matplotlib_available, seaborn_available
 from pyomo.contrib.pynumero.asl import AmplInterface
 from pyomo.opt import SolverFactory
-from pyomo.common.dependencies import attempt_import
-import os
 
 ipopt_available = SolverFactory("ipopt").available()
 pynumero_ASL_available = AmplInterface.available()
-
-pandas, pandas_available = attempt_import("pandas")
-matplotlib, matplotlib_available = attempt_import("matplotlib")
 
 
 @unittest.skipIf(
@@ -60,70 +55,6 @@ class TestRooneyBieglerExamples(unittest.TestCase):
         )
 
         likelihood_ratio_example.main()
-
-    # Test rooney biegler DOE example
-    @unittest.skipUnless(pandas_available, "test requires pandas")
-    @unittest.skipUnless(matplotlib_available, "test requires matplotlib")
-    def test_doe_example(self):
-        """
-        Tests the Design of Experiments (DoE) functionality, including
-        plotting logic, without displaying GUI windows.
-        """
-        import glob  # To handle multiple file matches
-        from pyomo.contrib.parmest.examples.rooney_biegler.doe_example import (
-            run_rooney_biegler_doe,
-        )
-
-        plt = matplotlib.pyplot
-
-        # Switch backend to 'Agg' to prevent GUI windows from popping up
-        plt.switch_backend('Agg')
-
-        # Define the prefix used to identify files to be cleaned up
-        file_prefix = "rooney_biegler"
-
-        # Ensure we clean up the file even if the test fails halfway through
-        def cleanup_file():
-            generated_files = glob.glob(f"{file_prefix}_*.png")
-            for f in generated_files:
-                try:
-                    os.remove(f)
-                except OSError:
-                    pass
-
-        self.addCleanup(cleanup_file)
-
-        # Run with plotting flags set to TRUE
-        results = run_rooney_biegler_doe(
-            optimize_experiment_A=True,
-            optimize_experiment_D=True,
-            compute_FIM_full_factorial=True,
-            draw_factorial_figure=True,
-            design_range={'hour': [0, 10, 3]},
-            plot_optimal_design=True,
-            tee=False,
-        )
-
-        # Assertions for Numerical Results
-        self.assertIn("D", results["optimization"])
-        self.assertIn("A", results["optimization"])
-
-        # Assertions for Plotting Logic
-        # Check that the custom plotting block ran and stored figures
-        self.assertTrue(
-            len(results["plots"]) > 0, "No plot objects were stored in results['plots']"
-        )
-
-        # Check that draw_factorial_figure actually created the file
-        expected_d_plot = f"{file_prefix}_D_opt.png"
-        self.assertTrue(
-            os.path.exists(expected_d_plot),
-            f"Expected plot file '{expected_d_plot}' was not created.",
-        )
-
-        # Explicitly close figures to free memory
-        for fig in results["plots"]:
-            plt.close(fig)
 
 
 @unittest.skipUnless(pynumero_ASL_available, "test requires libpynumero_ASL")
